+++ src/Cloud/DTO/CloudDatabase.php
@@ -4,14 +4,12 @@ declare(strict_types=1);
 
 namespace LaravelCloudBlueprint\Cloud\DTO;
 
-final readonly class CloudEnvironment
+final readonly class CloudDatabase
 {
     public function __construct(
         public string $id,
-        public string $applicationId,
+        public string $clusterId,
         public string $name,
-        public ?string $branch,
-        public ?string $databaseId = null,
     ) {
     }
 }