+++ src/Cloud/DTO/CloudUnknownDatabaseConfiguration.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LaravelCloudBlueprint\Cloud\DTO;
+
+final readonly class CloudUnknownDatabaseConfiguration implements CloudDatabaseClusterConfiguration
+{
+}